Revert "Move HeadersMap to types folder"
This reverts commit 3947a03e53f73028bafbedbbd9c5513e8d39e912.
