+++ src/Api/createRequest.ts
@@ -1,10 +1,9 @@
 import fetch from 'cross-fetch';
 
-import { HeadersMap } from '../types';
-
 import { createUrlWithQuery } from './lib';
-import { HttpCodes, Method, ApiResponse } from './types';
+import { HttpCodes, Method, HeadersMap, ApiResponse } from './types';
 import ApiError from './ApiError';
+
 import {
     CONTENT_TYPE,
     INVALID_URL_ERROR_MESSAGE,
