--- src/types/HeadersMap.ts
@@ -1,3 +1,0 @@
-export default interface HeadersMap {
-    [name: string]: string;
-}
